Updated the json format definition, changed terminology to LP.
Updating the json format definition to a more compact format.  Basically restructuring the
event data from labeled record objects to an array (removing the string name).  This
change reduces the size of the data files by more than an order of magnitude.

I also decided to change the terminology from "simulation object" to LP so that we conform
more directly the the standard terminology of the PDES community.