adding an analysis step to show how many LPs cover x% of the total events sent
basically i'm trying to capture an idea of the frequency of events sent between the
various LPs that one LP talks to.  so, for example, does the LP send one message to all
most of the LPs, but the bulk of its messages actually go to only a few LPs.  ideally this
number may help us to understand commucation patterns and look for opportunities for
partitioning guidance.  there are probably other (and better) metrics, this is just a
start.